requirements.txt
apache-airflow[gcp]
google-cloud-storage

apache-airflow-providers-google
google-auth
google-cloud-bigquery
google-api-python-client